requirements.txt
Flask==2.3.3
Flask_SQLAlchemy==3.1.1
Flask_Login==0.6.3
Flask_WTF==1.2.1
email_validator==2.1.1
gunicorn==21.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
flask-cors==4.0.0
pandas==2.2.2
pytz==2024.1
xlsxwriter==3.2.0
simple_salesforce==1.12.6
openpyxl==3.1.5
cryptography==43.0.0
Flask-Caching==2.1.0
schedule==1.2.1
psutil==5.9.8
requests==2.32.3
mailjet-rest==1.5.1
itsdangerous==2.2.0
